Change: Worker no longer has an explicit name
The name was a.) unnecessary and b.) misleading. When seeing a worker
name like "http/http", one would expect there to be other workers, which
were possibly disabled by the configuration. Since that is not the case,
the worker name was removed entirely. The score.serve module will just
call it 'http' now (or whatever it was renamed to in the configuration).